feat: S3 이미지 업로드, 수정, 삭제, 코드리뷰 수정

+++ ssfinder-be/src/main/java/com/ssfinder/domain/founditem/service/FoundItemService.java
@@ -147,6 +147,10 @@ public class FoundItemService {
         FoundItem foundItem = foundItemRepository.findById(foundId)
                 .orElseThrow(() -> new CustomException(ErrorCode.FOUND_ITEM_NOT_FOUND));
 
+        if(!foundItem.getUser().getId().equals(userId)) {
+            throw new CustomException(ErrorCode.FOUND_ITEM_ACCESS_DENIED);
+        }
+
         if (Objects.nonNull(foundItem.getImage())) {
             s3Service.deleteFile(foundItem.getImage());
         }
